+++ MotorBike/Models/Inspection.cs
@@ -15,7 +15,7 @@ public partial class Inspection
 
     public int ModelId { get; set; }
 
-    public byte YearNo { get; set; }
+    public short YearNo { get; set; }
 
     public string ChassisNo { get; set; } = null!;
